requirements.txt
Flask
tensorflow
Pillow
requests
numpy
gunicorn # for production deployment later